backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
fastapi==0.135.3
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-ws==0.9.0
idna==3.11
jiter==0.14.0
lmstudio==1.5.0
msgspec==0.21.0
oauthlib==3.3.1
ollama==0.6.1
openai==2.31.0
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
wsproto==1.3.2
